requirements.txt
bcrypt==4.3.0
cffi==1.17.1
cryptography==45.0.4
future==1.0.0
markdown-it-py==3.0.0
mdurl==0.1.2
netmiko==4.6.0
ntc_templates==7.9.0
paramiko==3.5.1
pycparser==2.22
Pygments==2.19.2
PyNaCl==1.5.0
pyserial==3.5
python-dotenv==1.1.1
PyYAML==6.0.2
rich==14.0.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
scp==0.15.0
six==1.17.0
textfsm==1.1.3
